DICOM parameters now have DICOM: prefix again, and XNATFileMatch moved to the dicoms.py to be close to the extraction function
