+++ src/xnatuploader/xnatuploader.py
@@ -15,8 +15,8 @@ __version__ = version("xnatuploader")
 
 from openpyxl import load_workbook
 
-from xnatuploader.matcher import Matcher, XNATFileMatch
-from xnatuploader.dicoms import dicom_extractor
+from xnatuploader.matcher import Matcher
+from xnatuploader.dicoms import dicom_extractor, XNATFileMatch
 from xnatuploader.workbook import new_workbook, add_filesheet, load_config
 from xnatuploader.upload import Upload, trigger_pipelines
